ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
motor==3.3.1
pymongo==4.5.0
email-validator==2.1.0.post1
python-dotenv==1.0.0
tenacity==8.2.3
colorlog==6.8.0
slowapi==0.1.8
gunicorn==21.2.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
pytest-mock==3.12.0